// enums/RoutePaths.ts
export enum RoutePaths {
  SIGN_UP = '/sign-up',
  SIGN_IN = '/sign-in',
  HOME = '/',
  STAYS = '/stays',
  EVENTS = '/events',
  ABOUT = '/about',
  PROFILE = '/profile',
  BADGES = '/badges',
  LEADERBOARD = '/leaderboard',
  COMMUNITIES = '/communities',
  SAVED = '/saved',
  SETTINGS = '/settings',
  PROPERTY = '/property/:id',
  TRIPS = '/trips',
  NOTIFICATIONS = '/notifications',
  CHAT = '/chat/:id',
  EMAIL_CONFIRMATION = '/email-confirmation',
  USER_PROFILE = '/user/:id',
  SETTINGS_SECURITY = '/settings/security',
  SETTINGS_PROFILE = '/settings/profile',
  SETTINGS_PAYMENTS = '/settings/payments',
  CHATS = '/chats',
  ACHIEVEMENTS = '/achievements',
  SAILING = '/sailing',
  QUESTS = '/quests',
  PRIVACY_POLICY = '/privacy-policy',
  TERMS_OF_USE = '/terms-of-use',
  BUG_REPORTS = '/bug-reports',
  REFERRAL = '/referral-program',
  SUITCASES = '/suitcases',
}